code/agent-server/requirements.txt
livekit-agents[google]~=1.2
python-dotenv